Upstream version 3.9.1

# gpg: Unterschrift vom Sa 31 Jan 2015 15:33:39 CET mittels RSA-Schlüssel ID 513B51E4
# gpg: Korrekte Unterschrift von "Markus Koschany <apo@gambaru.de>" [uneingeschränkt]
# gpg:                     alias "Markus Koschany <markus@koschany.net>" [uneingeschränkt]

+++ source/shared_lib/sources/streflop/libm/flt-32/e_gammaf_r.cpp
@@ -46,14 +46,13 @@ __ieee754_gammaf_r (Simple x, int *signgamp)
       *signgamp = 0;
       return (x - x) / (x - x);
     }
-  if (hx == 0xff800000)
+  if ((u_int32_t)hx == 0xff800000)
     {
       /* x == -Inf.  According to ISO this is NaN.  */
       *signgamp = 0;
       return x - x;
     }
 
-  /* XXX FIXME.  */
   return __ieee754_expf (__ieee754_lgammaf_r (x, signgamp));
 }
 }
